test: add frontend utils tests
