+++ tests/web/frontend/conftest.py
@@ -3,7 +3,7 @@ import pytest
 from src.web.frontend.frontend_testing import FrontendTestRunner, MockDataGenerator
 from src.web.frontend.frontend_app import FrontendAppFactory
 from src.web.frontend.frontend_router import create_router_with_default_routes
-from src.web.frontend.ui_interactions import (
+from src.web.frontend.utils import (
     create_mock_component,
     create_mock_route,
     create_mock_navigation_state,
